extract interafce
